Handle single instance deployment of port binded tasks
It seems the automatic update of tasks when the service is modified
is not working fine in the case of a single instance with tasks
depending on an specific port (it cannot start the new one because
port is occupied by the old one).

Solve that explicitely stopping the tasks, deploying the service
and restarting them.